Client : Refactor AddButton, ConfirmButton.vue, CancelButton.vue replace
